Refactor explain to mom meme to generator
